feat: convert to dollar

+++ src/main/java/com/example/demo_for_showing_test_containers/bank/controller/GetBanksResponseBankJson.java
@@ -11,7 +11,7 @@ record GetBanksResponseBankJson(int value) {
     }
     static GetBanksResponseBankJson of(Bank bank) {
         return new GetBanksResponseBankJson(
-                bank.money().deposit()
+                bank.money().deposit().intValue()
         );
     }
 }
